# management/registry/openapi/openapi.yaml
basePath: /
definitions:
...
  handler.CreateLineRequest:
    properties:
      display_order:
        type: integer
      name:
        type: string
      note:
        type: string
      region:
        type: string
    type: object
  handler.CreateLineResponse:
    properties:
      id:
        type: integer
      token:
        type: string
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  handler.ResetTokenResponse:
    properties:
      token:
        type: string
    type: object
...
  handler.SetLineNodesRequest:
    properties:
      nodes:
        items:
          type: string
        type: array
    type: object
...
  handler.StatusResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
...
  handler.UpdateLineRequest:
    properties:
      display_order:
        type: integer
      enabled:
        type: boolean
      name:
        type: string
      note:
        type: string
      region:
        type: string
    type: object
...
  model.Line:
    properties:
      created_at:
        type: string
      display_order:
        type: integer
      enabled:
        type: boolean
      id:
        type: integer
      name:
        type: string
      node_count:
        type: integer
      note:
        type: string
      region:
        type: string
      token:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: Hive management plane API for nodes and subscriptions.
  title: Hive Node Registry API
  version: 0.1.0
paths:
...
  /admin/lines:
    get:
      description: 返回所有线路（含节点数量）
      operationId: AdminListLines
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Line'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: 获取线路列表
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: 创建新线路并生成订阅 token
      operationId: AdminCreateLine
      parameters:
      - description: 线路信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateLineRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CreateLineResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: 创建线路
      tags:
      - admin
  /admin/lines/{id}:
    delete:
      description: 删除指定线路
      operationId: AdminDeleteLine
      parameters:
      - description: 线路 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: 删除线路
      tags:
      - admin
    patch:
      consumes:
      - application/json
      description: 更新指定线路的属性
      operationId: AdminUpdateLine
      parameters:
      - description: 线路 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新字段
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateLineRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: 更新线路
      tags:
      - admin
  /admin/lines/{id}/nodes:
    get:
      description: 返回指定线路关联的节点 MAC 列表
      operationId: AdminGetLineNodes
      parameters:
      - description: 线路 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: 获取线路节点
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: 替换指定线路的节点列表
      operationId: AdminSetLineNodes
      parameters:
      - description: 线路 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 节点 MAC 列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.SetLineNodesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: 设置线路节点
      tags:
      - admin
  /admin/lines/{id}/reset-token:
    post:
      description: 为指定线路生成新的订阅 token
      operationId: AdminResetLineToken
      parameters:
      - description: 线路 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResetTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: 重置线路 token
      tags:
      - admin
...
  /l/{token}:
    get:
      description: 返回 base64 编码的 VLESS 订阅链接
      operationId: PublicLineVless
      parameters:
      - description: 线路 token
        in: path
        name: token
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: VLESS subscription
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: 获取线路 VLESS 订阅
      tags:
      - subscription
  /l/{token}/clash:
    get:
      description: 返回 Clash/Mihomo YAML 格式订阅配置
      operationId: PublicLineClash
      parameters:
      - description: 线路 token
        in: path
        name: token
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Clash subscription
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: 获取线路 Clash 订阅
      tags:
      - subscription
...
schemes:
- https
- http
securityDefinitions:
  AdminSessionCookie:
    description: '"HttpOnly session cookie for admin UI"'
    in: cookie
    name: hive_admin_session
    type: apiKey
  BearerAuth:
    description: '"Authorization: Bearer <API_SECRET>"'
    in: header
    name: Authorization
    type: apiKey
  CustomerSessionCookie:
    description: '"HttpOnly session cookie for customer portal"'
    in: cookie
    name: hive_customer_session
    type: apiKey
swagger: "2.0"